requirements.txt
Django>=5.2.6
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
google-auth>=2.35.0
requests>=2.32.3
django-cors-headers
